long parsers

+++ jsdynamic-typify/src/test/scala/Example.scala
@@ -24,7 +24,7 @@ object jsDynamicExample {
   case object Male extends Gender
   case object Female extends Gender
 
-  case class Person(email: String @@ Email, age: Int @@ Age, gender: Gender, session: Option[Int @@ SessId])
+  case class Person(email: String @@ Email, age: Long @@ Age, gender: Gender, session: Option[Int @@ SessId])
   case class UnsafePerson(email: String, age: Int)
 
   val typify = new Typify[String, js.Dynamic]
@@ -39,7 +39,7 @@ object jsDynamicExample {
     case "f" => Female.successNel[String]
     case x => s"Invalid gender $x".failureNel[Gender]
   })
-  implicit lazy val vAge = typify.validate[Int, Int @@ Age]((k: String, a: Int) =>
+  implicit lazy val vAge = typify.validate[Long, Long @@ Age]((k: String, a: Long) =>
     (a > 18).option(tag[Age](a)).toSuccessNel(s"${k} too young"))
   implicit lazy val sid = typify.validate[Option[Int], Option[Int @@ SessId]]((i: Option[Int]) =>
     i match {
@@ -68,9 +68,9 @@ object jsDynamicExample {
 
   @JSExport
   def partialValidatePerson(jsd: String, root: Seq[String] = Seq()):
-  ValidationNel[String, (String, Int) => Person] =
-    typify[(String @@ Email, Int @@ Age) => Person](js.JSON.parse(jsd), root)
-      .map(fn => (s: String, i: Int) => fn(tag[Email](s), tag[Age](i)))
+  ValidationNel[String, (String, Long) => Person] =
+    typify[(String @@ Email, Long @@ Age) => Person](js.JSON.parse(jsd), root)
+      .map(fn => (s: String, i: Long) => fn(tag[Email](s), tag[Age](i)))
 }
